+++ CHANGELOG.md
@@ -17,7 +17,11 @@
     * get\_ours(api, section) -> actually just calls get(api, section, api)
   * DueBadges now has a class method get(api, section)
   * Evening now has class methods:
-    * 
+    * get\_programme(api, section\_id, term\_id)
+    * create(api, parameters)
+  * Evening now has instance methods:
+    * update(api)
+    * get\_badge\_requirements(api, evening)
   * Grouping now has a class method get\_for\_section(api, section)
   * Member now has a class method to get\_for\_section(api, section\_id, term\_id)
   * Section now has class methods:
@@ -28,6 +32,8 @@
     * get(api, term\_id)
     * get\_for\_section(api, section\_id)
     * get\_current\_term\_for\_section(api, section\_id)
+    * create(api, parameters)
+  * Term now has instance method update(api)
 
 ## Version 0.0.27